+++ packages/orq-rc/src/models/operations/deploymentgetconfig.ts
@@ -1492,6 +1492,7 @@ export type DeploymentGetConfigEncodingFormat = ClosedEnum<
  * Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response.
  */
 export const DeploymentGetConfigReasoningEffort = {
+  Disable: "disable",
   Low: "low",
   Medium: "medium",
   High: "high",