+++ packages/orq-rc/src/models/operations/deployments.ts
@@ -199,6 +199,7 @@ export type DeploymentsEncodingFormat = ClosedEnum<
  * Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response.
  */
 export const DeploymentsReasoningEffort = {
+  Disable: "disable",
   Low: "low",
   Medium: "medium",
   High: "high",