+++ packages/orq-rc/src/models/operations/updateprompt.ts
@@ -99,6 +99,7 @@ export type UpdatePromptEncodingFormat = ClosedEnum<
  * Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response.
  */
 export const UpdatePromptReasoningEffort = {
+  Disable: "disable",
   Low: "low",
   Medium: "medium",
   High: "high",
@@ -568,6 +569,7 @@ export type UpdatePromptPromptsEncodingFormat = ClosedEnum<
  * Constrains effort on reasoning for reasoning models. Reducing reasoning effort can result in faster responses and fewer tokens used on reasoning in a response.
  */
 export const UpdatePromptPromptsReasoningEffort = {
+  Disable: "disable",
   Low: "low",
   Medium: "medium",
   High: "high",