+++ packages/orq-rc/RELEASES.md
@@ -2359,3 +2359,13 @@ Based on:
 - [typescript v3.9.0-rc.5] packages/orq-rc
 ### Releases
 - [NPM v3.9.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/3.9.0-rc.5 - packages/orq-rc
+
+## 2025-06-24 19:49:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.568.2 (2.634.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.9.0-rc.6] packages/orq-rc
+### Releases
+- [NPM v3.9.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/3.9.0-rc.6 - packages/orq-rc